refactor: move batch bookmark insertion logic to domains/bookmarks.ts
