+++ src/server.ts
@@ -5,7 +5,7 @@ import {
   deleteActiveToken,
 } from "./database.ts";
 import { z } from "zod";
-import { getBookmarkDataFromUrl } from "./domains/bookmarks.ts";
+import { getBookmarkDataFromUrl, insertManyBookmarks } from "./domains/bookmarks.ts";
 import { embedText } from "./ai/embeddings.ts";
 import { getLoggerConfig } from "./logger.ts";
 import type { BookmarkWithContent } from "./types.ts";
@@ -63,47 +63,12 @@ server.post("/bookmarks/batch", {
     const batchBodySchema = z.array(bodySchema);
     const body = batchBodySchema.parse(request.body);
 
-    const bookmarks: BookmarkWithContent[] = [];
-    const BATCH_SIZE = 20;
-    let totalProcessed = 0;
-    let totalSuccess = 0;
-    let totalFailed = 0;
-
-    for (let i = 0; i < body.length; i += BATCH_SIZE) {
-      const batch = body.slice(i, i + BATCH_SIZE);
-      const batchResults = await Promise.allSettled(
-        batch.map((item) => getBookmarkDataFromUrl(item.url)),
-      );
-
-      batchResults.forEach((result, index) => {
-        const url = batch[index].url;
-        totalProcessed++;
-
-        if (result.status === "fulfilled") {
-          bookmarks.push(result.value);
-          totalSuccess++;
-        } else {
-          request.log.error(
-            `Failed to process bookmark for URL: ${url}`,
-            result.reason,
-          );
-          totalFailed++;
-        }
-      });
-    }
-
-    if (bookmarks.length > 0) {
-      await insertBookmarks(bookmarks);
-    }
+    const urls = body.map((item) => item.url);
+    const stats = await insertManyBookmarks(urls);
 
     return {
       success: true,
-      stats: {
-        totalProcessed,
-        totalSuccess,
-        totalFailed,
-        bookmarksInserted: bookmarks.length,
-      },
+      stats,
     };
   },
 });
